feat(si): decoded accessors for frequency_list, local_time_offset, PDC
Completes the decoded-accessor audit (issues #37/#38 + the broader BCD/packed
fields):
- FrequencyListDescriptor::centre_frequencies_hz() / set_centre_frequencies_hz()
  (per coding_type; satellite decodes like the satellite delivery descriptor)
- LocalTimeOffsetEntry::local_time_offset()/next_time_offset()/time_of_change()
  + set_offsets()/set_time_of_change() (signed chrono::Duration; the two offsets
  share one polarity bit, so set_offsets requires matching signs)
- PdcDescriptor::pil_day/month/hour/minute() + set_pil()

Also: drop redundant `as` casts in dvb-common, fix redundant-explicit doc links,
rustfmt. Round-trip tested in dvb-si/tests/accessors.rs (11 cases).

+++ dvb-common/src/bcd.rs
@@ -41,7 +41,7 @@ pub fn bcd_to_decimal(raw: u64, nibbles: u8) -> Option<u64> {
     }
     let mut acc = 0u64;
     for i in (0..nibbles).rev() {
-        let digit = ((raw >> (i * 4)) & 0x0F) as u64;
+        let digit = (raw >> (i * 4)) & 0x0F;
         if digit > 9 {
             return None;
         }
